+++ CHANGELOG.md
@@ -9,6 +9,8 @@ selinux Cookbook CHANGELOG
 - Resolved cookstyle errors
 - Update package information for debian based on https://debian-handbook.info/browse/stable/sect.selinux.html
  - selinux-activate looks like it's required to ACTUALLY activate selinux on non-RHEL systems. This seems like it could be destructive if unexpected.
+- Add property temporary to allow for switching between permissive and enabled
+- Add install resource
 
 v0.9.0 (2015-02-22)
 -------------------